// types/schema.ts
export interface components {
  schemas: {
// ...
    Billing: {
      /** @description Personal details of the customer needed for billing. */
      name?: string;
      organization?: components['schemas']['Organization'];
      address?: components['schemas']['Address'];
      /** Format: email */
      email?: string;
      phone?: string;
      time?: components['schemas']['Time'];
      tax_number?: string;
      /** Format: date-time */
      created_at?: string;
      /** Format: date-time */
      updated_at?: string;
    };
// ...
    Context: {
      domain: components['schemas']['Domain'];
      country: components['schemas']['Country']['code'];
      city: components['schemas']['City']['code'];
      /**
       * @description Defines the Beckn API call. Any actions other than the enumerated actions are not supported by Beckn Protocol
       * @enum {string}
       */
      action:
      | 'search'
      | 'select'
      | 'init'
      | 'confirm'
      | 'update'
      | 'status'
      | 'track'
      | 'cancel'
      | 'rating'
      | 'support'
      | 'on_search'
      | 'on_select'
      | 'on_init'
      | 'on_confirm'
      | 'on_update'
      | 'on_status'
      | 'on_track'
      | 'on_cancel'
      | 'on_rating'
      | 'on_support';
      /** @description Version of Beckn core API specification being used */
      core_version: string;
      /**
       * Format: uri
       * @description Unique id of the BAP. By default it is the fully qualified domain name of the BAP
       */
      bap_id: string;
      /**
       * Format: uri
       * @description URI of the BAP for accepting callbacks. Must have the same domain name as the bap_id
       */
      bap_uri: string;
      /**
       * Format: uri
       * @description Unique id of the BPP. By default it is the fully qualified domain name of the BPP
       */
      bpp_id?: string;
      /**
       * Format: uri
       * @description URI of the BPP. Must have the same domain name as the bap_id
       */
      bpp_uri?: string;
      /** @description This is a unique value which persists across all API calls from search through confirm */
      transaction_id: string;
      /** @description This is a unique value which persists during a request / callback cycle */
      message_id: string;
      /**
       * Format: date-time
       * @description Time of request generation in RFC3339 format
       */
      timestamp: string;
      /** @description The encryption public key of the sender */
      key?: string;
      /** @description The duration in ISO8601 format after timestamp for which this message holds valid */
      ttl?: string;
    };
// ...
    Fulfillment: {
      /** @description Unique reference ID to the fulfillment of an order */
      id?: string;
      /** @description This describes the type of fulfillment */
      type?: string;
      provider_id?: components['schemas']['Provider']['id'];
      rating?: components['schemas']['Rating']['value'];
      state?: components['schemas']['State'];
      /**
       * @description Indicates whether the fulfillment allows tracking
       * @default false
       */
      tracking?: boolean;
      customer?: {
        person?: components['schemas']['Person'];
        contact?: components['schemas']['Contact'];
      };
      agent?: components['schemas']['Agent'];
      person?: components['schemas']['Person'];
      contact?: components['schemas']['Contact'];
      vehicle?: components['schemas']['Vehicle'];
      /** @description Details on the start of fulfillment */
      start?: {
        location?: components['schemas']['Location'];
        time?: components['schemas']['Time'];
        instructions?: components['schemas']['Descriptor'];
        contact?: components['schemas']['Contact'];
        person?: components['schemas']['Person'];
        authorization?: components['schemas']['Authorization'];
      };
      /** @description Details on the end of fulfillment */
      end?: {
        location?: components['schemas']['Location'];
        time?: components['schemas']['Time'];
        instructions?: components['schemas']['Descriptor'];
        contact?: components['schemas']['Contact'];
        person?: components['schemas']['Person'];
        authorization?: components['schemas']['Authorization'];
      };
      rateable?: components['schemas']['Rateable'];
      tags?: components['schemas']['Tags'][];
      stops?:components['schemas']['Stop'][];
    };
// ...
    Intent: {
      descriptor?: components['schemas']['Descriptor'];
      provider?: components['schemas']['Provider'];
      fulfillment?: components['schemas']['Fulfillment'];
      payment?: components['schemas']['Payment'];
      category?: components['schemas']['Category'];
      offer?: components['schemas']['Offer'];
      item?: components['schemas']['Item'];
      tags?: components['schemas']['Tags'];
    };
// ...
    Order: {
// ...
      fulfillments?: components['schemas']['Fulfillment'][];
      return_terms?: components['schemas']['ReturnTerms'];
      update_terms?: components['schemas']['UpdateTerms'];
      quote?: components['schemas']['Quotation'];
      payment?: components['schemas']['Payment'];
      /** Format: date-time */
      created_at?: string;
      /** Format: date-time */
      updated_at?: string;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}